+++ composeApp/src/wasmJsMain/kotlin/tech/kotlinlang/permission/main.kt
@@ -1,7 +1,6 @@
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import kotlinx.browser.document
-import kotlinx.browser.window
 import tech.kotlinlang.permission.App
 
 @OptIn(ExperimentalComposeUiApi::class)