+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
-- Add logic to check camera permission in WasmJs
+- Add logic to check camera permission in WasmJs in Permission module
+- Add logic to ask and check Notification Permission in Permission module
 
 ---